# (perf): Improve database update on 'immediate