--- a/org-roam-db.el
+++ b/org-roam-db.el
@@ -518,13 +518,8 @@ If FORCE, force a rebuild of the cache from scratch."
          (org-agenda-files nil)
          (org-roam-files (org-roam--list-all-files))
          (current-files (org-roam-db--get-current-files))
-         (id-count 0)
-         (link-count 0)
-         (tag-count 0)
-         (title-count 0)
-         (ref-count 0)
+         (count-plist nil)
          (deleted-count 0)
-         (modified-count 0)
          (modified-files nil))
     (dolist (file org-roam-files)
       (let ((contents-hash (org-roam-db--file-hash file)))
@@ -536,8 +531,33 @@ If FORCE, force a rebuild of the cache from scratch."
         ;; These files are no longer around, remove from cache...
         (org-roam-db--clear-file file)
         (setq deleted-count (1+ deleted-count)))
+    (setq count-plist (org-roam-db--update-files modified-files (length org-roam-files)))
+    (org-roam-message "total: Δ%s, files-modified: Δ%s, ids: Δ%s, links: Δ%s, tags: Δ%s, titles: Δ%s, refs: Δ%s, deleted: Δ%s"
+                      (plist-get count-plist :total-count)
+                      (plist-get count-plist :modified-count)
+                      (plist-get count-plist :id-count)
+                      (plist-get count-plist :link-count)
+                      (plist-get count-plist :tag-count)
+                      (plist-get count-plist :title-count)
+                      (plist-get count-plist :ref-count)
+                      deleted-count)))
+
+(defun org-roam-db--update-files (modified-files total-count)
+  (let* ((gc-cons-threshold org-roam-db-gc-threshold)
+         (org-agenda-files nil)
+         (error-count 0)
+         (id-count 0)
+         (link-count 0)
+         (tag-count 0)
+         (title-count 0)
+         (ref-count 0)
+         (deleted-count 0)
+         (modified-count 0)
+ )
+    ;; Clear database entry for file
     (pcase-dolist (`(,file . _) modified-files)
       (org-roam-db--clear-file file))
+
     ;; Process all the files for IDs first
     ;;
     ;; We do this so that link extraction is cheaper: this eliminates the need
@@ -555,10 +575,13 @@ If FORCE, force a rebuild of the cache from scratch."
               (when org-roam-enable-headline-linking
                 (setq id-count (+ id-count (org-roam-db--insert-ids)))))
           (file-error
-           (setq org-roam-files (remove file org-roam-files))
+           (setq total-count (1- total-count))
+           (setq error-count (1+ error-count))
            (org-roam-db--clear-file file)
            (lwarn '(org-roam) :warning
                   "Skipping unreadable file while building cache: %s" file)))))
+
+    ;; Process titles, tags, links and ref links of file
     (pcase-dolist (`(,file . _) modified-files)
       (org-roam-message "Processed %s/%s modified files..." modified-count (length modified-files))
       (condition-case nil
@@ -569,19 +592,12 @@ If FORCE, force a rebuild of the cache from scratch."
               (setq title-count (+ title-count (org-roam-db--insert-titles)))
               (setq ref-count (+ ref-count (org-roam-db--insert-refs))))
           (file-error
-           (setq org-roam-files (remove file org-roam-files))
+           (setq total-count (1- total-count))
+           (setq error-count (1+ error-count))
            (org-roam-db--clear-file file)
            (lwarn '(org-roam) :warning
                   "Skipping unreadable file while building cache: %s" file))))
-    (org-roam-message "total: Δ%s, files-modified: Δ%s, ids: Δ%s, links: Δ%s, tags: Δ%s, titles: Δ%s, refs: Δ%s, deleted: Δ%s"
-                      (length org-roam-files)
-                      modified-count
-                      id-count
-                      link-count
-                      tag-count
-                      title-count
-                      ref-count
-                      deleted-count)))
+    (list :total-count total-count :error-count error-count :modified-count modified-count :id-count id-count :title-count title-count :tag-count tag-count :link-count link-count :ref-count ref-count)))
 
 (defun org-roam-db-update ()
   "Update the database."