--- a/org-roam-db.el
+++ b/org-roam-db.el
@@ -549,21 +549,19 @@ If FORCE, force a rebuild of the cache from scratch."
   (caar (org-roam-db-query [:select hash :from files
                               :where (= file $s1)] file-path)))
 
-(defun org-roam-db-update-file (&optional file-path)
+(defun org-roam-db-update-file (file-path)
   "Update Org-roam cache for FILE-PATH.
 If the file does not exist anymore, remove it from the cache.
 If the file exists, update the cache with information."
-  (setq file-path (or file-path
-                      (buffer-file-name (buffer-base-buffer))))
   (let ((content-hash (org-roam-db--file-hash file-path))
         (db-hash  (org-roam-db--get-file-hash-from-db file-path)))
     (unless (string= content-hash db-hash)
       (org-roam-db--update-files (list (cons file-path content-hash)))
-      (org-roam-message "updated db for: %s" file-path))))
+      (org-roam-message "Updated: %s" file-path))))
 
 (defun org-roam-db--update-files (modified-files)
   "Update Org-roam cache for a list of MODIFIED-FILES.
-Expects a list of file, hash pair."
+FILES is a list of (file . hash) pairs."
   (let* ((gc-cons-threshold org-roam-db-gc-threshold)
          (org-agenda-files nil)
          (error-count 0)
@@ -573,10 +571,8 @@ Expects a list of file, hash pair."
          (title-count 0)
          (ref-count 0)
          (modified-count 0))
-    ;; Clear database entry for file
     (pcase-dolist (`(,file . _) modified-files)
       (org-roam-db--clear-file file))
-
     ;; Process all the files for IDs first
     ;;
     ;; We do this so that link extraction is cheaper: this eliminates the need
@@ -603,24 +599,24 @@ Expects a list of file, hash pair."
     (pcase-dolist (`(,file . _) modified-files)
       (org-roam-message "Processed %s/%s modified files..." modified-count (length modified-files))
       (condition-case nil
-            (org-roam--with-temp-buffer file
-              (setq modified-count (1+ modified-count))
-              (setq link-count (+ link-count (org-roam-db--insert-links)))
-              (setq tag-count (+ tag-count (org-roam-db--insert-tags)))
-              (setq title-count (+ title-count (org-roam-db--insert-titles)))
-              (setq ref-count (+ ref-count (org-roam-db--insert-refs))))
-          (file-error
-           (setq error-count (1+ error-count))
-           (org-roam-db--clear-file file)
-           (lwarn '(org-roam) :warning
-                  "Skipping unreadable file while building cache: %s" file))))
+          (org-roam--with-temp-buffer file
+            (setq modified-count (1+ modified-count))
+            (setq link-count (+ link-count (org-roam-db--insert-links)))
+            (setq tag-count (+ tag-count (org-roam-db--insert-tags)))
+            (setq title-count (+ title-count (org-roam-db--insert-titles)))
+            (setq ref-count (+ ref-count (org-roam-db--insert-refs))))
+        (file-error
+         (setq error-count (1+ error-count))
+         (org-roam-db--clear-file file)
+         (lwarn '(org-roam) :warning
+                "Skipping unreadable file while building cache: %s" file))))
     (list :error-count error-count :modified-count modified-count :id-count id-count :title-count title-count :tag-count tag-count :link-count link-count :ref-count ref-count)))
 
 (defun org-roam-db-update ()
   "Update the database."
   (pcase org-roam-db-update-method
     ('immediate
-     (org-roam-db-update-file))
+     (org-roam-db-update-file (buffer-file-name (buffer-base-buffer))))
     ('idle-timer
      (org-roam-db-mark-dirty))
     (_