#1212 slowed down database updates considerably for me. I have 1493 files in my org roam directory. `(org-roam--list-all-files)` is too slow (2.4s) even on `rg` which slows down `org-roam-db-build-cache` considerably.

I understand the advantage of #1212 is that it allows for a single code path. Therefore, I extracted the core logic from `org-roam-db-build-cache` to allow applying database update on a list of files under a single code path.

I have tested it to be working on 'immediate method of database update but wouldn't be hard to back propagate stuff from previous commits that made `org-roam-db--update-file` work on a timer.

The following are benchmarks done on my largest file (184KB, 4820 lines).
`org-roam-db-build-cache` is the defacto database update mechanism.
`org-roam-db--update-file` is the old update mechanism that was deprecated in #1212.
`org-roam-db-update-file` is what this PR is attempting to introduce.

runtime without modifying file.
``` elisp
~/org-roam:master*? λ (benchmark 1 '(org-roam-db-build-cache))
Elapsed time: 2.757124s (0.189366s in 1 GCs)
~/org-roam:master*? λ (benchmark 1 '(org-roam-db--update-file file-path))
Elapsed time: 1.001571s (0.200564s in 1 GCs)
~/org-roam:master*? λ (benchmark 1 '(org-roam-db-update-file file-path))
Elapsed time: 0.008179s
```

run time after modifying the file.
``` elisp
~/org-roam:master*? λ (benchmark 1 '(org-roam-db-build-cache))
Elapsed time: 3.365744s
~/org-roam:master*? λ (benchmark 1 '(org-roam-db--update-file file-path))
Elapsed time: 0.987487s (0.212442s in 1 GCs)
~/org-roam:master*? λ (benchmark 1 '(org-roam-db-update-file file-path))
Elapsed ime: 0.895598s
```

database queries and listing all files
``` elisp
~/org-roam:master*? λ (benchmark 1 '(org-roam-db--get-file-hash-from-db file-path))
Elapsed time: 0.002355s
~/org-roam:master*? λ (benchmark 1 '(org-roam-db--get-current-files))
Elapsed time: 0.034420s
~/org-roam:master*? λ (benchmark 1 '(org-roam--list-all-files))
Elapsed time: 2.746063s (0.353111s in 1 GCs)
```

I did a couple of tests, and the run time of `org-roam-db-build-cache` is variable. For example, the runtime of  `org-roam-db-build-cache` increases to what is reported above after rebuilding the database.
``` elisp
~ λ (benchmark 1 '(org-roam-db-build-cache))
Elapsed time: 0.340561s
~ λ (benchmark 1 '(org-roam-db-build-cache))
Elapsed time: 1.293372s (0.172139s in 1 GCs)
```

| function                   | runtime on unmodified | runtime on modified |
|----------------------------|-----------------------|---------------------|
| `org-roam-db-build-cache`  | 2.757s                | 3.365s              |
| `org-roam-db-build-cache` fast  | 0.350s                | 1.298s              |
| `org-roam-db--update-file` | 1.001s                | 0.987s              |
| `org-roam-db-update-file`  | **0.008s**            | **0.895s**          |