@alphapapa 2020-11-25
FYI, you should generally run benchmarks compiled and with lexical-binding.  See macros here for example: https://github.com/alphapapa/emacs-package-dev-handbook#optimization