@natask 2020-11-25
@alphapapa  Interesting. 
I ran a couple of tests with `bench`, `bench-multi-lexical` and `bench-multi` from the link you provided but saw no discernible difference.
``` elisp
/u/s/e/28.0.50/lisp/eshell λ (benchmark 1 '(org-roam-db-update-file file-name))
Elapsed time: 0.007320s
/u/s/e/28.0.50/lisp/eshell λ (benchmark 1 '(org-roam-db-update-file file-name))
Elapsed time: 1.059315s (0.197132s in 1 GCs)
/u/s/e/28.0.50/lisp/eshell λ (benchmark 1 '(org-roam-db-update-file file-name))
Elapsed time: 0.874786s
/u/s/e/28.0.50/lisp/eshell λ (benchmark 1 '(org-roam-db-update-file file-name))
Elapsed time: 1.100952s (0.203795s in 1 GCs)
/u/s/e/28.0.50/lisp/eshell λ (bench 1 '(org-roam-db-update-file file-name))
(("Total runtime" "# of GCs" "Total GC runtime")
 hline
 (3.061e-06 0 0.0))
/u/s/e/28.0.50/lisp/eshell λ (bench 1 (org-roam-db-update-file file-name))
(("Total runtime" "# of GCs" "Total GC runtime")
 hline
 (0.846945444 0 0.0))
/u/s/e/28.0.50/lisp/eshell λ (bench 155 (org-roam-db-update-file file-name))
(("Total runtime" "# of GCs" "Total GC runtime")
 hline
 (0.177032377 0 0.0))
/u/s/e/28.0.50/lisp/eshell λ (bench 155 (org-roam-db-update-file file-name))
(("Total runtime" "# of GCs" "Total GC runtime")
 hline
 (1.073752271 0 0.0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.846481" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.002467" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.894440" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "2.408341" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.879444" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.875127" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.003398" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.800721" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi-lexical :ensure-equal t
  :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.004063" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi-lexical :ensure-equal t
  :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.002846" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi-lexical :ensure-equal t
  :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.897810" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi-lexical :ensure-equal t
  :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.002988" 0 0))
/u/s/e/28.0.50/lisp/eshell λ (bench-multi-lexical :ensure-equal t
  :forms (("org-roam-db" (org-roam-db-update-file file-name))))
(("Form" "x faster than next" "Total runtime" "# of GCs" "Total GC runtime")
 hline
 ("org-roam-db" "slowest" "0.826260" 0 0))

```
And I didn't notice discernible difference between the numbers I reported here.

I believe elisp libraries are compiled by default (most libraries I have seen have .elc file as well). Are elisp functions by default executed with lexical-binding?