@alphapapa 2020-11-25
> I believe elisp libraries are compiled by default (most libraries I have seen have .elc file as well). 

If they are installed with a package manager that compiles them, or if the user compiles them, yes.

> Are elisp functions by default executed with lexical-binding?

Please see the documentation regarding lexical binding in the link I posted.

> Is it possible to migrate these benefits to day to day function calls?

I don't know what you mean.  Maybe you want to look into byte-compiling your init file.